Preparation Phase: Use a mutable class to introduce a mutability bug.
When calculating the shipment, the approval date also changes.

A few other  problems were also introduced in ShipmentService:
It also changes the Order status and return the new shipment date,
it is a query and command method at the same time.

And most important, given it is the place where it is mutating the
Order status, it also requires the Mock to simulate it for the test to pass.
If the same mutation is not included in the production code we have a bug.
The classic case of the mock shadowing the production code.